MoveComponent adds the current movement state to the player object
