+++ Assets/Scripts/Movement/MoveStates/CharacterState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: db4753c3c53533d499c9c3bf9e2cc10e
+guid: d757c6ddf42970f46b92bc9b2c767cb6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
